% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{as_binary_data}
\alias{as_binary_data}
\title{Coerce user data into required binary wide format: n x (2d)}
\usage{
as_binary_data(
  x,
  format = c("auto", "wide", "long"),
  times = NULL,
  id_col = "snp",
  ind_col = "ind",
  time_col = "time",
  value_col = "y"
)
}
\arguments{
\item{x}{matrix/data.frame}

\item{format}{"auto" | "wide" | "long"}

\item{times}{optional numeric vector length d}

\item{id_col, }{ind_col, time_col, value_col used when long format}
}
\value{
list(y, d_single, times_single)
}
\description{
Supported:
\itemize{
\item wide matrix/data.frame: n x (2d), ordered as ind1(t1..td), ind2(t1..td)
\item long data.frame: columns for id, ind, time, value (ind must have exactly 2 levels)
}
}
