% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{as_eval_df}
\alias{as_eval_df}
\title{Collect evaluation files produced by user scripts (optional helper)}
\usage{
as_eval_df(eval_list)
}
\arguments{
\item{eval_list}{list of evaluation objects OR data.frame}
}
\value{
data.frame with J and BIC
}
\description{
Collect evaluation files produced by user scripts (optional helper)
}
\examples{
as_eval_df(data.frame(J = 2:4, BIC = c(120, 100, 110), loglik = c(-50, -40, -42)))
}
