% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_sad_dist.R
\name{.mcmcgraph_expose_sad_to_global}
\alias{.mcmcgraph_expose_sad_to_global}
\title{Ensure SAD distribution functions are visible to nimble}
\usage{
.mcmcgraph_expose_sad_to_global()
}
\description{
Ensure SAD distribution functions are visible to nimble
}
\keyword{internal}
