% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_sad_dist.R
\name{.mcmcgraph_register_sad}
\alias{.mcmcgraph_register_sad}
\title{Register SAD distribution in nimble (once per session)}
\usage{
.mcmcgraph_register_sad(force = FALSE)
}
\description{
Register SAD distribution in nimble (once per session)
}
\keyword{internal}
