% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_eval.R
\name{eval_bic}
\alias{eval_bic}
\title{Compute BIC from a fitted object (posterior mean plug-in)}
\usage{
eval_bic(fit, x, format = c("auto", "wide", "long"))
}
\arguments{
\item{fit}{mcmcgraph_result}

\item{x}{original input data (same used in fitting)}

\item{format}{"auto" | "wide" | "long"}
}
\value{
data.frame with J, loglik, BIC
}
\description{
Compute BIC from a fitted object (posterior mean plug-in)
}
