% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot_trace_density}
\alias{plot_trace_density}
\title{Plot trace + density for selected parameters}
\usage{
plot_trace_density(result, params = c("^phi", "^p\\\\["), max_params = 6)
}
\arguments{
\item{result}{mcmcgraph_result from run_mcmc_binary()}

\item{params}{character vector of parameter names (regex allowed)}

\item{max_params}{max number of params to plot}
}
\description{
Plot trace + density for selected parameters
}
