% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bic_eval.R
\name{fit_many_J}
\alias{fit_many_J}
\title{Fit multiple J values and return BIC table}
\usage{
fit_many_J(x, J_grid = 2:6, ...)
}
\arguments{
\item{x}{input data (wide/long)}

\item{J_grid}{integer vector of candidate J}

\item{...}{passed to run_mcmc_binary (e.g., niter, thin)}
}
\value{
list(fits, eval) where eval is data.frame(J,BIC,loglik)
}
\description{
Fit multiple J values and return BIC table
}
\examples{
\donttest{
data(example_binary)
res <- fit_many_J(example_binary$y, J_grid = 2:4,
                  times = example_binary$times, niter = 300)
res$eval
}
}
